+++ tests/app/controller/command/commands/mention_test.py
@@ -21,14 +21,7 @@ class MentionCommandTeset(TestCase):
     def test_handle_no_input(self):
         """Test handle command with no additional args."""
         self.assertEqual(self.testcommand.handle('UFJ42EU67', user),
-                         (self.testcommand.help, 200))
-
-    def test_handle_help(self):
-        """Test handle command help."""
-        self.assertEqual(
-            self.testcommand.handle(('help'), user),
-            (self.testcommand.help, 200)
-        )
+                         ("not supported", 200))
 
     def test_handle_add_karma_to_another_user(self):
         """Test handle command with karma to another user."""
@@ -37,6 +30,7 @@ class MentionCommandTeset(TestCase):
         self.mock_facade.retrieve.return_value = user
         self.assertEqual(self.testcommand.handle(('UFJ42EU67 ++'), user),
                          ("gave 1 karma to U123456789", 200))
+        self.mock_facade.retrieve.assert_called_once_with(User, "UFJ42EU67")
 
     def test_handle_add_karma_to_self(self):
         """Test handle command with karma to self."""
